[feat] 버킷수정하기 (1)
내용 수정 + 친구 추가/삭제 기능
